Add files via upload
The Rain Alert Project is a Python-based application developed to notify users when rainfall is expected in their area. It uses the OpenWeatherMap API to fetch real-time weather data and analyze precipitation probability. It is also very important to use the latitude and longitude of your location to ensure the code works accurately and efficiently. When rain is forecasted, the program automatically sends an alert message to the user via Twilio, ensuring timely updates through WhatsApp, email, or SMS. The automation of this program is handled using PythonAnywhere, allowing it to run continuously without manual execution. This project enhanced my understanding of API integration, JSON data parsing, and Python-based automation. (Note: When using an upgraded PythonAnywhere account, environment variables should be used to securely store API keys and authentication tokens.)